% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clonal_Family_Inference.R
\name{BCR.cluster}
\alias{BCR.cluster}
\title{Function: The core step of fastBCR to infer clonal families from processed data.}
\usage{
BCR.cluster(input, cluster_thre = 3, overlap_thre = 0.1, consensus_thre = 0.8)
}
\arguments{
\item{input}{AIRR format data after Data_Processing}

\item{cluster_thre}{Minimal clustering criteria. Defaluts to 3. For high efficiency, the threshold is increased by 1 for every 100,000 entries of input data.}

\item{overlap_thre}{The overlap coefficient threshold for merging two clusters whose selectable range is (0,1). Defaluts to 0.1. Lower thresholds may lead to overclustering while higher may lead to split of clonal families.}

\item{consensus_thre}{The consensus score threshold for filtering candidates. Defaluts to 0.8. A higher threshold means stricter inference of the cluster.}
}
\value{
Clonal families inferred by fastBCR
}
\description{
BCR clonal family inference consists of two steps, fast k-mer-based pre-clustering and optimized clustering, to infer clonal families fast and accurately.
}
