% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcr_clusters}
\alias{bcr_clusters}
\title{Data: As a sample dataset for Evolutionary_Analysis.}
\format{
An object of class \code{list} of length 827.
}
\usage{
bcr_clusters
}
\description{
Inferred clonal familes by fastBCR (using 'BCR.cluster(input)').
}
\examples{
data("bcr_clusters")
}
\keyword{datasets}
