% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clonal_Family_Simulation.R
\name{CF2fas}
\alias{CF2fas}
\title{Function: Generation of a fasta file consisting of sequences from distinct clonal familes.}
\usage{
CF2fas(germline_data, CF_n, mut_ratio, filename)
}
\arguments{
\item{germline_data}{Annotated germline sequences.}

\item{CF_n}{The number of simulated clonal families.}

\item{mut_ratio}{The point mutation rate for SHM during simulated affinity maturation process whose selectable range is (0,1).}

\item{filename}{Path to save fasta file.}
}
\value{
A saved fasta file consisting of sequences from distinct clonal familes that need to be annotated subsequently.
}
\description{
Function: Generation of a fasta file consisting of sequences from distinct clonal familes.
}
