% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{clonal.tree.plot}
\alias{clonal.tree.plot}
\title{Plot: B cell lineage tree reconstructed by ClonalTree}
\usage{
clonal.tree.plot(nk_path)
}
\arguments{
\item{nk_path}{Path to the ClonalFamily_index.nk (the reconstructed BCR lineage tree in newick format)}
}
\value{
the reconstructed BCR lineage tree
}
\description{
Orange points represent nodes and blue points represent tips
The x-axis shows the absolute genetic distance
}
