% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{Clonal.tree}
\alias{Clonal.tree}
\title{Function: Reconstructing B cell lineage trees with minimum spanning tree and genotype abundances (ClonalTree)}
\usage{
Clonal.tree(bcr_clusters, n)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{n}{select a cluster to plot the evolutionary tree}
}
\value{
the reconstructed BCR lineage tree
}
\description{
Function: Reconstructing B cell lineage trees with minimum spanning tree and genotype abundances (ClonalTree)
}
\examples{
data("bcr_clusters")
Clonal.tree(bcr_clusters, 1)
}
