% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{clu.size.df}
\alias{clu.size.df}
\title{Function: Calculate the size and number of clusters for the samples within each group}
\usage{
clu.size.df(clusters_list, group_label)
}
\arguments{
\item{clusters_list}{A list where each element of the list is the clonal families inferred by fastBCR for each sample}

\item{group_label}{Label of group}
}
\value{
Dataframe containing information on the group, cluster size, and number
}
\description{
Function: Calculate the size and number of clusters for the samples within each group
}
