% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{clu2df}
\alias{clu2df}
\title{Function: Merge clustered sequences from all BCR clusters and remove duplicates}
\usage{
clu2df(bcr_cluster)
}
\arguments{
\item{bcr_cluster}{clonal families inferred by fastBCR}
}
\value{
clustered sequences from all BCR clusters
}
\description{
Function: Merge clustered sequences from all BCR clusters and remove duplicates
}
