% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{Clustered.seqs}
\alias{Clustered.seqs}
\title{Function: Classification of clustered and unclustered sequences}
\usage{
Clustered.seqs(pro_data_list, clusters_list)
}
\arguments{
\item{pro_data_list}{A list where each element is the preprocessed data named after its filename}

\item{clusters_list}{A list where each element is a list of clonal families inferred by fastBCR from a single sample}
}
\value{
A list containing clustered and unclustered sequences in each sample
'clustered_seqs' is a list where each element is the clustered sequences in a sample named after its filename
'unclustered_seqs' is a list where each element is the unclustered sequences in a sample named after its filename
}
\description{
Function: Classification of clustered and unclustered sequences
}
