% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{Clusters.gene.fre}
\alias{Clusters.gene.fre}
\title{Function: Calculate the frequency of gene usage in clustered sequences from the samples}
\usage{
Clusters.gene.fre(clustered_seqs_sample, colname, uni_gene, group_label)
}
\arguments{
\item{clustered_seqs_sample}{A list where each element is the clustered sequences for each sample}

\item{colname}{'v_call' for V gene or 'j_call' for J gene}

\item{uni_gene}{All unique genes in two groups of samples}

\item{group_label}{Label of group}
}
\value{
Dataframe containing the gene usage frequency for samples
}
\description{
Function: Calculate the frequency of gene usage in clustered sequences from the samples
}
