% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{Clusters.summary}
\alias{Clusters.summary}
\title{Function: Summarize the number of clusters, the average size of clusters and the proportion of clustered sequences}
\usage{
Clusters.summary(pro_data_list, clusters_list)
}
\arguments{
\item{pro_data_list}{A list where each element is the preprocessed data named after its filename}

\item{clusters_list}{A list where each element is a list of clonal families inferred by fastBCR from a single sample}
}
\value{
A list where each element is the summary of clusters from a single sample
}
\description{
Function: Summarize the number of clusters, the average size of clusters and the proportion of clustered sequences
}
