% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{Clusters.visualization}
\alias{Clusters.visualization}
\title{Plot: Visualization of clusters from a sample}
\usage{
Clusters.visualization(pro_data_list, clusters_list, index)
}
\arguments{
\item{pro_data_list}{A list where each element is the preprocessed data named after its filename}

\item{clusters_list}{A list where each element is a list of clonal families inferred by fastBCR from a single sample}

\item{index}{The Index of sample for visualization}
}
\value{
Point diagram of clusters from a sample where a circle represents a cluster. The size and color of the circle represents the size of the cluster.
}
\description{
Plot: Visualization of clusters from a sample
}
