% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clonal_Family_Inference.R
\name{consensus_scores}
\alias{consensus_scores}
\title{Function: Calculate consensus scores of clonal families infered by fastBCR}
\usage{
consensus_scores(bcr_clusters)
}
\arguments{
\item{bcr_clusters}{Clonal families infered by fastBCR}
}
\value{
A list where each element is the consensus score of each bcr_cluster
}
\description{
Function: Calculate consensus scores of clonal families infered by fastBCR
}
