% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{CSR.sample}
\alias{CSR.sample}
\title{Function: Network of CSR within a sample}
\usage{
CSR.sample(bcr_clusters)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}
}
\value{
Visualization of isotype co-occurrence within BCR clusters. Circle size represents the number of clusters carrying a given isotype. Lines connecting two circles indicate the enrichment level of observing switches in the two corresponding immunoglobulin subclasses. The enrichment level is the ratio of observed and expected switches if immunoglobulin isotypes are assumed to be independently distributed among clusters.
}
\description{
Plot CSR network within a sample.
}
\examples{
data("bcr_clusters")
CSR.sample(bcr_clusters)
}
