% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{data.load}
\alias{data.load}
\title{Function: Load datasets into a list}
\usage{
data.load(folder_path, storage_format)
}
\arguments{
\item{folder_path}{Path to the data folder}

\item{storage_format}{The storage format of data. It can be in 'csv', 'tsv', or 'Rdata' format.}
}
\value{
A list where each element is the raw data named after its filename
}
\description{
The compressed files in '7zip', 'cab', 'cpio', 'iso9660', 'lha', 'mtree', 'shar', 'rar', 'raw', 'tar', 'tar.gz', 'xar', 'zip', 'warc' format can also be read in
}
