% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{data.preprocess}
\alias{data.preprocess}
\title{Function: Preprocessing of raw data to meet the input requirements for clonal family inference}
\usage{
data.preprocess(data_list)
}
\arguments{
\item{data_list}{A list where each element is the raw data named after its filename}
}
\value{
A list where each element is the preprocessed data named after its filename
}
\description{
Function: Preprocessing of raw data to meet the input requirements for clonal family inference
}
