% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBCR-package.R
\docType{package}
\name{fastBCR-package}
\alias{fastBCR}
\alias{fastBCR-package}
\title{fastBCR: a k-mer-based approach for fast inferencing clonal families from large-scale B cell repertoire sequencing data}
\description{
fastBCR is an automatic BCR clonal family inference method, which also incorporates multiple functional modules for downstream analyses. fastBCR is composed of two parts: Fast BCR clonal family inference & Downstream analysis.
}
\keyword{internal}
