% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBCR-p.R
\name{filter_public_antibodies}
\alias{filter_public_antibodies}
\title{Filter Public Antibodies by Threshold}
\usage{
filter_public_antibodies(prediction_results, threshold = 0.5)
}
\arguments{
\item{prediction_results}{A data frame containing prediction results, including a `public_score` column.}

\item{threshold}{A numeric value specifying the public score threshold for classification. Defaults to 0.5.}
}
\value{
A list containing two data frames:
  - `high_public`: Rows where `public_score` is greater than or equal to the threshold.
  - `low_public`: Rows where `public_score` is less than the threshold.
}
\description{
This function filters predicted public antibody results into high and low public antibody categories based on a threshold.
}
