% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{gene.fre.df}
\alias{gene.fre.df}
\title{Calculation: Gene usage frequency of clustered sequences from a sample}
\usage{
gene.fre.df(clustered_seqs, colname, uni_gene, group_label)
}
\arguments{
\item{clustered_seqs}{Clustered sequences from a sample}

\item{colname}{'v_call' for V gene or 'j_call' for J gene}

\item{uni_gene}{All unique genes in two groups of samples}

\item{group_label}{Label of group}
}
\value{
Data.frame for gene usage frequency
}
\description{
Calculation: Gene usage frequency of clustered sequences from a sample
}
