% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{gene.fre.plot}
\alias{gene.fre.plot}
\title{Plot: Boxplot of the gene usage frequency between the two groups.}
\usage{
gene.fre.plot(df)
}
\arguments{
\item{df}{Data.frame for the gene usage frequency between the two groups}
}
\value{
Boxplot
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
