% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clonal_Family_Simulation.R
\name{germline2fas}
\alias{germline2fas}
\title{Function: Generation of a fasta file consisting of germline sequences}
\usage{
germline2fas(family_num)
}
\arguments{
\item{family_num}{The number of simulated clonal families.}
}
\value{
A faste file of randomly generated germline sequences.
}
\description{
Randomly generate germline sequences and convert them to a fasta file.
}
\examples{
data("ighv_hum_df")
data("ighd_hum_df")
data("ighj_hum_df")
germline_fasta = germline2fas(10)
}
