% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{germline_data}
\alias{germline_data}
\title{Data: As a sample data for Simulation}
\format{
An object of class \code{data.frame} with 47 rows and 48 columns.
}
\usage{
germline_data
}
\description{
Data: As a sample data for Simulation
}
\examples{
data("germline_data")
}
\keyword{datasets}
