% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ighv_hum_df}
\alias{ighv_hum_df}
\title{ighv_hum_df}
\format{
A data frame with 119 rows and 2 variables:
\describe{
  \item{gene}{The gene name}
  \item{seq}{The corresponding sequence}
}
}
\source{
IMGT
}
\usage{
ighv_hum_df
}
\description{
human germline IgH-V (heavy chain v gene segments). When multiple alleles were present, the
first one was included. These names and sequences can be changed by
customized by changing this dataframe. Additionally, repeating elements
can give certain germline gene elements a larger probability of being used
during repertoire evolution.
}
\keyword{datasets}
