% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input}
\alias{input}
\title{Data: As a sample dataset for BCR_clustering.}
\format{
An object of class \code{data.frame} with 20906 rows and 164 columns.
}
\usage{
input
}
\description{
Processed data after Data_Processing (using 'data.pro(test_data)').
}
\examples{
data("input")
bcr_clusters = BCR.cluster(input)
}
\keyword{datasets}
