% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{junc.len}
\alias{junc.len}
\title{Function: Histogram of junction length}
\usage{
junc.len(data, colname)
}
\arguments{
\item{data}{BCR repertoire}

\item{colname}{"DNA" for junction length or "AA" for junction_aa length}
}
\value{
junction (DNA/AA) length distribution
}
\description{
Basic statistics of repertoire junction length distribution.
}
\examples{
data("input")
junc.len(input, "AA")
}
