% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{len.clustered.plot}
\alias{len.clustered.plot}
\title{Plot: Density ridges of junction amino acid length between clustered and unclustered sequence
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test}
\usage{
len.clustered.plot(clustered_seqs, unclustered_seqs)
}
\arguments{
\item{clustered_seqs}{The clustered sequences}

\item{unclustered_seqs}{The unclustered sequences}
}
\value{
junction amino acid length distribution between the two groups
}
\description{
Plot: Density ridges of junction amino acid length between clustered and unclustered sequence
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
