% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{len.group.plot}
\alias{len.group.plot}
\title{Plot: Density ridges of junction amino acid length between two groups
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test}
\usage{
len.group.plot(
  group1_all_clustered_seqs,
  group1_label,
  group2_all_clustered_seqs,
  group2_label
)
}
\arguments{
\item{group1_all_clustered_seqs}{All the clustered sequences in group1}

\item{group1_label}{Label of group1}

\item{group2_all_clustered_seqs}{All the clustered sequences in group2}

\item{group2_label}{Label of group2}
}
\value{
junction amino acid length distribution between the two groups
}
\description{
Plot: Density ridges of junction amino acid length between two groups
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
