% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{len.sample.plot}
\alias{len.sample.plot}
\title{Plot: Histogram and density plot of junction amino acid length of clustered sequences from a sample}
\usage{
len.sample.plot(clustered_seqs)
}
\arguments{
\item{clustered_seqs}{Clustered sequences from a sample}
}
\value{
junction amino acid length distribution from a sample
}
\description{
Plot: Histogram and density plot of junction amino acid length of clustered sequences from a sample
}
