% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{Metrics.df}
\alias{Metrics.df}
\title{Calculation: Data.frame for metrics of clusters}
\usage{
Metrics.df(input, bcr_clusters, group_label)
}
\arguments{
\item{input}{BCR repertoire}

\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{group_label}{Label used to divide different groups}
}
\value{
Data.frame for metrics of bcr clusters, including Tcf20 score, average SHM ratio of clusters and SHM ratios of four isotypes (IGHD, IGHM, IGHA, IGHG).
}
\description{
Calculation: Data.frame for metrics of clusters
}
