% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{msa.plot}
\alias{msa.plot}
\title{Plot: Visualization of multiple sequence alignment (MSA) of junction sequences within a cluster}
\usage{
msa.plot(bcr_clusters, index, type = c("AA", "DNA"), raw_data = NA)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{index}{Index of cluster}

\item{type}{'DNA' for deoxyribonucleic acid or 'AA' for amino acid}

\item{raw_data}{The raw data which the clonal families inferred from
It is needed if you want to plot the MSA of 'DNA' sequences.
fastBCR will retrieve all the DNA sequences, which can be multiple sequences due to the degeneracy of codons, that correspond to the amino acid sequence of each clonotype from the raw data}
}
\value{
Visualization of MSA. From top to bottom are title('Vgene_Jgene_Length'), seqlogo, ggmsa, and msabar
}
\description{
Perform MSA on the junction of all sequences in the clonal family for visualization
}
