% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{msa.tree}
\alias{msa.tree}
\title{Function: Plot of evolutionary tree and multiple sequences alignment (MSA)}
\usage{
msa.tree(bcr_clusters, n)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{n}{select a cluster to plot the evolutionary tree}
}
\value{
evolutionary tree & MSA
}
\description{
Perform MSA on all sequences in the clonal family and plot an evolutionary tree based on Hamming distance.
}
\examples{
data("bcr_clusters")
msa.tree(bcr_clusters, 20)
}
