% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{NAb.query}
\alias{NAb.query}
\title{Function: Query the corresponding sequence from the public antibody database}
\usage{
NAb.query(bcr_clusters, AbDab, method = NA, maxDist = NA, species = "Human")
}
\arguments{
\item{bcr_clusters}{Clonal families inferred by fastBCR}

\item{AbDab}{The public antibody database}

\item{method}{CDRH3 matching method. It can be 'NA' for perfect match, 'hamming' for hamming distance or 'lv' for Levenshtein distance. Defaults to 'NA'.}

\item{maxDist}{Maximum distance allowed for matching when the argument 'method' is 'hamming' or 'lv'. Defaults to 'NA'.}

\item{species}{can be 'Mouse' or 'Human'. Defaults to 'Human'.}
}
\value{
Result of the query
}
\description{
Function: Query the corresponding sequence from the public antibody database
}
