% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{NAb.ratio.calculation}
\alias{NAb.ratio.calculation}
\title{Function: Calculate the NAb ratios between the two groups.}
\usage{
NAb.ratio.calculation(
  pro_data_list1,
  clusters_list1,
  group1_label,
  pro_data_list2,
  clusters_list2,
  group2_label,
  AbDab,
  method = NA,
  maxDist = NA,
  species = "Human"
)
}
\arguments{
\item{pro_data_list1}{A list where each element is the preprocessed data named after its filename in group1}

\item{clusters_list1}{A list where each element of the list is the clonal families inferred by fastBCR for each sample in group1}

\item{group1_label}{Label of group1}

\item{pro_data_list2}{A list where each element is the preprocessed data named after its filename in group2}

\item{clusters_list2}{A list where each element of the list is the clonal families inferred by fastBCR for each sample in group2}

\item{group2_label}{Label of group2}

\item{NAb_vjcdr3}{IGHV gene, IGHJ gene and CDRH3 of all NAb sequences}
}
\value{
Dataframe containing the NAb ratios between the two groups.
}
\description{
NAb ratio is established as an indicator of the proportional prevalence of neutralizing antibody sequences within expanded clonotypes in each sample
It is defined as the fraction of the number of NAb sequences within clonal families to the total number of NAb sequences present in each sample
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
