% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{NAb.ratio.plot}
\alias{NAb.ratio.plot}
\title{Plot: Boxplot of the NAb ratios between the two groups.}
\usage{
NAb.ratio.plot(df)
}
\arguments{
\item{df}{Data.frame for NAb ratios of samples}
}
\value{
Boxplot
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
