% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{NAb.ratio}
\alias{NAb.ratio}
\title{Calculation: NAb ratio (the ratio derived from dividing the tally of NAb sequences encompassed by clonal families by the aggregate count of NAb sequences present within each sample)}
\usage{
NAb.ratio(input, bcr_clusters, NAb_v, NAb_j, NAb_cdr3)
}
\arguments{
\item{input}{BCR repertoire}

\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{NAb_v}{V gene of all NAb sequences}

\item{NAb_j}{J gene of all NAb sequences}

\item{NAb_cdr3}{CDR3 sequences of all NAb sequences}
}
\value{
NAb ratio
}
\description{
Calculation: NAb ratio (the ratio derived from dividing the tally of NAb sequences encompassed by clonal families by the aggregate count of NAb sequences present within each sample)
}
