% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{NAb.roc.plot}
\alias{NAb.roc.plot}
\title{Plot: ROC curve of the NAb ratios between the two groups.}
\usage{
NAb.roc.plot(roc)
}
\arguments{
\item{roc}{a “roc” object, a list of class “roc” buliding from pROC package}
}
\value{
ROC curve
}
\description{
Plot: ROC curve of the NAb ratios between the two groups.
}
