% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBCR-p.R
\name{paired.preprocess}
\alias{paired.preprocess}
\title{Function: Preprocess datasets and check required columns of paired data}
\usage{
paired.preprocess(paired_raw_data_list)
}
\arguments{
\item{paired_raw_data_list}{A list of paired data.}
}
\value{
Processed data as input for clonal family inference
}
\description{
To infer clonal families successfully, the dataset should include essential columns:
1. v_call_heavy: heavy chain V gene with or without allele,
2. j_call_heavy: heavy chain J gene with or without allele,
3. junction_aa_heavy: amino acid translation of the heavy chain junction,
4. v_call_light: light chain V gene with or without allele,
5. j_call_light: light chain J gene with or without allele,
6. junction_aa_light: amino acid translation of the light chain junction
}
