% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{pie.freq.plot}
\alias{pie.freq.plot}
\title{Plot: Pieplot of V/J gene usage frequency}
\usage{
pie.freq.plot(clustered_seqs, colname = c("v_call", "j_call"))
}
\arguments{
\item{clustered_seqs}{Clustered sequences}

\item{colname}{'v_call' for V gene or 'j_call' for J gene}
}
\value{
Pie chart. The top ten most frequent genes are shown, and the rest are represented by 'others'
}
\description{
Plot: Pieplot of V/J gene usage frequency
}
