% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{pie.freq}
\alias{pie.freq}
\title{Function: Pieplot of V/J usage frequency}
\usage{
pie.freq(data, colname)
}
\arguments{
\item{data}{BCR repertoire}

\item{colname}{"v_call" for V gene or "j_call" for J gene}
}
\value{
pie chart (The top ten most frequent genes are shown, and the rest are represented by 'others')
}
\description{
Basic statistics of repertoire gene segment usage.
}
\examples{
data("input")
pie.freq(input, "v_call")
}
