% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.clonal.tree}
\alias{plot.clonal.tree}
\title{Plot: Reconstructing B cell lineage trees with minimum spanning tree and genotype abundances (ClonalTree)}
\usage{
\method{plot}{clonal.tree}(bcr_clusters, index)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{index}{index of cluster}
}
\value{
the reconstructed BCR lineage tree
}
\description{
Plot: Reconstructing B cell lineage trees with minimum spanning tree and genotype abundances (ClonalTree)
}
