% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.clu.size}
\alias{plot.clu.size}
\title{Plot: Bubble plot between the size and number of clusters between the two groups}
\usage{
\method{plot}{clu.size}(df)
}
\arguments{
\item{df}{Data.frame for the size and number of clusters between the two groups}
}
\value{
Bubble plot
}
\description{
Plot: Bubble plot between the size and number of clusters between the two groups
}
