% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.CSR.cluster}
\alias{plot.CSR.cluster}
\title{Plot: Network of CSR within a cluster}
\usage{
\method{plot}{CSR.cluster}(bcr_clusters, index)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{index}{index of cluster}
}
\value{
Visualization of isotype co-occurrence within a BCR cluster. Circle size represents the number of sequences carrying a given isotype. Lines connecting two circles indicate the enrichment level of observing switches in the two corresponding immunoglobulin subclasses. The enrichment level is the ratio of observed and expected switches if immunoglobulin isotypes are assumed to be independently distributed among cluster.
}
\description{
Plot: Network of CSR within a cluster
}
