% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.len.group}
\alias{plot.len.group}
\title{Plot: Density ridges of junction amino acid length}
\usage{
\method{plot}{len.group}(df)
}
\arguments{
\item{df}{Data.frame for the junction amino acid length of clustered sequences between the two groups}
}
\value{
junction amino acid length distribution between the two groups
}
\description{
Plot: Density ridges of junction amino acid length
}
