% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.len.sample}
\alias{plot.len.sample}
\title{Plot: Histogram and density plot of junction amino acid length of clustered sequences from a sample}
\usage{
\method{plot}{len.sample}(clustered_seqs)
}
\arguments{
\item{clustered_seqs}{Clustered sequences from a sample}
}
\value{
junction amino acid length distribution from a sample
}
\description{
Plot: Histogram and density plot of junction amino acid length of clustered sequences from a sample
}
