% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.NAb.ratio}
\alias{plot.NAb.ratio}
\title{Plot: Boxplot of the NAb ratios between the two groups.}
\usage{
\method{plot}{NAb.ratio}(df)
}
\arguments{
\item{df}{Data.frame for NAb ratios of samples}
}
\value{
Boxplot
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
