% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.NAb.roc}
\alias{plot.NAb.roc}
\title{Plot: ROC curve of the NAb ratios between the two groups.}
\usage{
\method{plot}{NAb.roc}(roc)
}
\arguments{
\item{roc}{a “roc” object, a list of class “roc” buliding from pROC package}
}
\value{
ROC curve
}
\description{
Plot: ROC curve of the NAb ratios between the two groups.
}
