% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.Tcf20}
\alias{plot.Tcf20}
\title{Plot: Boxplot of the Tcf20 scores between the two groups.}
\usage{
\method{plot}{Tcf20}(df)
}
\arguments{
\item{df}{Data.frame for the Tcf20 score of clusters between the two groups}
}
\value{
Boxplot
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
