% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{plot.treemap}
\alias{plot.treemap}
\title{Plot: Tree map of clusters}
\usage{
\method{plot}{treemap}(input, bcr_clusters)
}
\arguments{
\item{input}{BCR repertoire}

\item{bcr_clusters}{clonal families inferred by fastBCR}
}
\value{
tree map
}
\description{
Plot: Tree map of clusters
}
