% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBCR-p.R
\name{predict_public_antibody}
\alias{predict_public_antibody}
\title{Predict Public Antibody Scores}
\usage{
predict_public_antibody(data, model = "cdrh", python_env = "public")
}
\arguments{
\item{data}{A data frame containing BCR data. Must include columns: `cdr1`, `cdr2`, `cdr3`, and `vgene`.}

\item{model}{A character string specifying the prediction model to use. Options are:
- `"cdrh"`: Model trained on heavy chain CDR regions.
- `"cdrl"`: Model trained on light chain CDR regions.
- `"cdrh3"`: Model trained on heavy chain CDR3 region only.
- `"cdrl3"`: Model trained on light chain CDR3 region only.}

\item{python_env}{A character string specifying the Python environment to use. Defaults to `"public"`.}
}
\value{
A data frame with an additional column, `public_score`, containing the predicted public antibody scores.
}
\description{
This function predicts the public antibody scores for input BCR sequences using pre-trained Python models.
}
