% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{seqlogo.plot}
\alias{seqlogo.plot}
\title{Plot: Visualization of sequence logo of junction sequences within a cluster}
\usage{
seqlogo.plot(bcr_clusters, index, type = c("AA", "DNA"))
}
\arguments{
\item{bcr_clusters}{Clonal families inferred by fastBCR}

\item{index}{Index of cluster}

\item{type}{'DNA' or 'AA'}
}
\value{
Sequence logo
}
\description{
Perform sequence logo on the junction of all sequences in the clonal family for visualization
}
