% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{seqlogo}
\alias{seqlogo}
\title{Function: Visualization of seqlogo of junction_aa sequences within a cluster}
\usage{
seqlogo(bcr_clusters, n)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{n}{select a cluster to visualize}
}
\value{
cluster sequences visualization. From top to bottom are title(Vgene_Jgene_junction_aaLength), seqlogo, ggmsa, and msabar
}
\description{
Perform seqlogo on the junction amino acid of all sequences in the clonal family for visualization.
}
\examples{
data("bcr_clusters")
seqlogo(bcr_clusters, 20)
}
