% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{SHM.clu}
\alias{SHM.clu}
\title{Calculation: SHM ratios of clusters from a sample}
\usage{
SHM.clu(bcr_clusters)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}
}
\value{
SHM ratios of clusters
}
\description{
Calculation: SHM ratios of clusters from a sample
}
