% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{SHM.cluster}
\alias{SHM.cluster}
\title{Function: Scatter diagram of cluster size (x axis), cluster number (y axis) and SHM ratio (point size) of all clusters in a sample}
\usage{
SHM.cluster(bcr_clusters)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}
}
\value{
A scatter plot of cluster size and cluster number of all clusters inferred from a repertoire. The size of point is positively correlated with the SHM ratio.
}
\description{
Plot scatter diagram of all clusters in a sample. Each point represents clusters with a given clusters size.
}
\examples{
data("bcr_clusters")
SHM.cluster(bcr_clusters)
}
