% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{SHM.iso.plot}
\alias{SHM.iso.plot}
\title{Plot: Boxplot of SHM ratios of clustered sequences in different isotypes}
\usage{
SHM.iso.plot(SHM_iso_df)
}
\arguments{
\item{SHM_df}{Dataframe containing the average SHM ratio of all clusters and the SHM ratios of clustered sequences in four isotypes in each sample}
}
\value{
Boxplot showing the SHM ratios of clustered sequences in four isotypes ('IGHD', 'IGHM', 'IGHA', and 'IGHG') between two groups
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
