% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{SHM.iso}
\alias{SHM.iso}
\title{Calculation: SHM ratios of clustered sequences in different isotypes from a sample}
\usage{
SHM.iso(bcr_clusters)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}
}
\value{
SHM ratios of clustered sequences in four isotypes ("IGHD", "IGHM", "IGHA", and "IGHG")
}
\description{
Calculation: SHM ratios of clustered sequences in different isotypes from a sample
}
