% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{SHM.plot}
\alias{SHM.plot}
\title{Plot: Boxplot of the SHM ratios between the two groups.}
\usage{
SHM.plot(df)
}
\arguments{
\item{df}{Data.frame for the SHM ratios of clusters between the two groups}
}
\value{
Boxplot
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
