% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SHM_ratio_2}
\alias{SHM_ratio_2}
\title{Data: As a sample data for plot.SHM.sample (sample2).}
\format{
An object of class \code{numeric} of length 4.
}
\usage{
SHM_ratio_2
}
\description{
Example2 for SHM ratios of four isotypes (using 'SHM.sample(bcr_clusters_2)').
}
\examples{
data("SHM_ratio_2")
}
\keyword{datasets}
