% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{Tcf20.plot}
\alias{Tcf20.plot}
\title{Plot: Boxplot of the Tcf20 scores between the two groups.}
\usage{
Tcf20.plot(clusters_list1, group1_label, clusters_list2, group2_label)
}
\arguments{
\item{clusters_list1}{A list where each element of the list is the clonal families inferred by fastBCR for each sample in group1}

\item{group1_label}{Label of group1}

\item{clusters_list2}{A list where each element of the list is the clonal families inferred by fastBCR for each sample in group2}

\item{group2_label}{Label of group2}
}
\value{
Boxplot showing the Tcf20 scores between the two groups
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
