% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Data: As a sample dataset for Data_Processing.}
\format{
An object of class \code{data.frame} with 44839 rows and 164 columns.
}
\usage{
test_data
}
\description{
A real experimental BCR sequencing repertoire with AIRR format.
}
\examples{
data("test_data")
input = data.pro(test_data)
}
\keyword{datasets}
