% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{vjpair.group.plot}
\alias{vjpair.group.plot}
\title{Plot: Heatmap of the fold change of V-J gene pair frequency between groups}
\usage{
vjpair.group.plot(
  group1_seqs_list,
  group1_all_clustered_seqs,
  group1_label,
  group2_seqs_list,
  group2_all_clustered_seqs,
  group2_label
)
}
\arguments{
\item{group1_seqs_list}{A list containing clustered and unclustered sequences in each sample in group1}

\item{group1_all_clustered_seqs}{All the clustered sequences in group1}

\item{group1_label}{Label of group1}

\item{group2_seqs_list}{A list containing clustered and unclustered sequences in each sample in group2}

\item{group2_all_clustered_seqs}{All the clustered sequences in group2}

\item{group2_label}{Label of group2}
}
\value{
Heatmap showing the fold change of V-J gene pair frequency of clustered sequences between two groups.
}
\description{
Log fold change (log FC) is calculated as the log2 ratio of the average values between group1 and group2 samples
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
FDR correction was performed with the Benjamini–Hochberg procedure
}
