% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCR_Clustering.R
\name{BCR.cluster}
\alias{BCR.cluster}
\title{Function: The core step of fastBCR to infer clonal families from processed data.}
\usage{
BCR.cluster(
  input,
  similarity_thre = 0.1,
  cluster_thre = 3 + floor(nrow(input)/1e+05),
  compactness_thre = 0.8
)
}
\arguments{
\item{input}{AIRR format data after Data_Processing}

\item{similarity_thre}{The similarity threshold for merging two clusters whose selectable range is (0,1]. Defaluts to 0.1. Lower thresholds may lead to overclustering while higher may lead to split of clonal families.}

\item{cluster_thre}{Minimal clustering criteria. Defaluts to (3 + floor(nrow(input)/100000).}

\item{compactness_thre}{The compactness threshold for filtering clusters. Defaluts to 0.8. A higher threshold means stricter inference of the cluster.}
}
\value{
Clonal families inferred by fastBCR
}
\description{
BCR Clustering consists of two steps, k-mer-based pre-clustering and optimized clustering, to infer clonal families fast and accurately.
}
\examples{
data("input")
bcr_clusters <- BCR.cluster(input)
}
