% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{clu.size.plot}
\alias{clu.size.plot}
\title{Plot: Bubble plot between the size and number of clusters between the two groups}
\usage{
clu.size.plot(df)
}
\arguments{
\item{df}{Data.frame for the size and number of clusters between the two groups}
}
\value{
Bubble plot
}
\description{
Plot: Bubble plot between the size and number of clusters between the two groups
}
