% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Processing.R
\name{data.preprocess}
\alias{data.preprocess}
\title{Function: Preprocessing of raw data to meet the input requirements for clonal family inference}
\usage{
data.preprocess(data_list, count_col_name = NA)
}
\arguments{
\item{data_list}{A list where each element is the raw data named after its filename}

\item{count_col_name}{The column name for the count of each sequence.
It can be "consensus_count", "duplicate_count" or "umi_count" according to your needs.
Defaults to "NA" which means the original count of the sequence is not taken into account.}
}
\value{
A list where each element is the processed data named after its filename
}
\description{
Function: Preprocessing of raw data to meet the input requirements for clonal family inference
}
