% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Processing.R
\name{data.pro}
\alias{data.pro}
\title{Function: The first step of fastBCR to process raw data.}
\usage{
data.pro(raw_data)
}
\arguments{
\item{raw_data}{AIRR format data. Inference of clonal familes requires following columns to be present in the raw_data: "sequence_id", "v_call", "j_call", "junction_aa". "junction" or "c_call" are optional if you want to plot "evolutionary tree" or need isotypes related analyis (SHM/CSR).}
}
\value{
Processed data as input for BCR_Clustering
}
\description{
The input of the function needs to meet the AIRR standard format (containing at least 'sequence_id', 'v_call', 'j_call', and 'junction_aa' information). Only productive sequences whose junction amino acid lengths between 9 and 26 are reserved. Sequences with the same 'v_call', 'j_call' and 'junction_aa' are considered identical and deduplicated.
}
\examples{
data("test_data")
input <- data.pro(test_data)
}
