% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{gene.fre.plot}
\alias{gene.fre.plot}
\title{Plot: Boxplot of the gene usage frequency between the two groups.}
\usage{
gene.fre.plot(
  group1_seqs_list,
  group1_all_clustered_seqs,
  group1_label,
  group2_seqs_list,
  group2_all_clustered_seqs,
  group2_label,
  colname
)
}
\arguments{
\item{group1_seqs_list}{A list containing clustered and unclustered sequences in each sample in group1}

\item{group1_all_clustered_seqs}{All the clustered sequences in group1}

\item{group1_label}{Label of group1}

\item{group2_seqs_list}{A list containing clustered and unclustered sequences in each sample in group2}

\item{group2_all_clustered_seqs}{All the clustered sequences in group2}

\item{group2_label}{Label of group2}

\item{colname}{'v_call' for V gene or 'j_call' for J gene}
}
\value{
Boxplot showing the V/J gene usage of the clustered sequences between the two groups.
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
