% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clonal_Family_Simulation.R
\name{germline2fas}
\alias{germline2fas}
\title{Function: Generation of a fasta file consisting of germline sequences.}
\usage{
germline2fas(gemline_num, filename)
}
\arguments{
\item{gemline_num}{The number of simulated germline sequences.}

\item{filename}{Path to save fasta file.}
}
\value{
A saved fasta file consisting of randomly generated germline sequences that need to be annotated subsequently.
}
\description{
Randomly generate germline sequences based in V(D)J recombination and convert them to a fasta file.
}
\examples{
germline2fas(100, filename = "Simulation/Germline.fasta")
}
