% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{len.group.plot}
\alias{len.group.plot}
\title{Plot: Density ridges of junction amino acid length}
\usage{
len.group.plot(df)
}
\arguments{
\item{df}{Data.frame for the junction amino acid length of clustered sequences between the two groups}
}
\value{
junction amino acid length distribution between the two groups
}
\description{
Plot: Density ridges of junction amino acid length
}
