% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{msa.plot}
\alias{msa.plot}
\title{Plot: Visualization of multiple sequence alignment (MSA) of junction sequences within a cluster}
\usage{
msa.plot(bcr_clusters, index, type = c("AA", "DNA"))
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{index}{index of cluster}

\item{type}{'DNA' or 'AA'}
}
\value{
Visualization of MSA. From top to bottom are title("Vgene_Jgene_Length"), seqlogo, ggmsa, and msabar
}
\description{
Perform MSA on the junction of all sequences in the clonal family for visualization
}
