% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{seqlogo.plot}
\alias{seqlogo.plot}
\title{Plot: Visualization of seqlogo of junction sequences within a cluster}
\usage{
seqlogo.plot(bcr_clusters, index, type = c("AA", "DNA"))
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}

\item{index}{index of cluster}

\item{type}{'DNA' or 'AA'}
}
\value{
Seqlogo
}
\description{
Perform seqlogo on the junction of all sequences in the clonal family for visualization
}
