% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{SHM.iso.plot}
\alias{SHM.iso.plot}
\title{Plot: Boxplot of SHM ratios of clustered sequences in different isotypes from a group}
\usage{
SHM.iso.plot(df)
}
\arguments{
\item{df}{Data.frame for SHM ratios of clustered sequences in four isotypes ("IGHD", "IGHM", "IGHA", and "IGHG") from multiple samples}
}
\value{
Boxplot
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
