% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{SHM.iso}
\alias{SHM.iso}
\title{Function: Calculation of SHM ratios in different isotypes in each sample}
\usage{
SHM.iso(bcr_clusters)
}
\arguments{
\item{bcr_clusters}{clonal families inferred by fastBCR}
}
\value{
four SHM ratio in different isotypes ("IGHD", "IGHM", "IGHA", and "IGHG") calculated by all clusters inferred from a repertoire
}
\description{
Calculate SHM ratios in different isotypes of all clusters inferred from a repertoire.
}
\examples{
data("bcr_clusters")
SHM.iso(bcr_clusters)
}
