% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{SHM.plot}
\alias{SHM.plot}
\title{Plot: Boxplot of the SHM ratios between the two groups}
\usage{
SHM.plot(SHM_df)
}
\arguments{
\item{SHM_df}{Dataframe containing the average SHM ratio of all clusters and the SHM ratios of clustered sequences in four isotypes in each sample}
}
\value{
Boxplot showing the SHM ratios between the two groups
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
