% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{Tcf20.plot}
\alias{Tcf20.plot}
\title{Plot: Boxplot of the Tcf20 scores between the two groups.}
\usage{
Tcf20.plot(df)
}
\arguments{
\item{df}{Data.frame for the Tcf20 score of clusters between the two groups}
}
\value{
Boxplot
}
\description{
Statistical comparisons are carried out by the two-sided Wilcoxon rank-sum test
}
