% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{vjpair.group.plot}
\alias{vjpair.group.plot}
\title{Plot: Heatmap of VJpair gene usage frequency between groups}
\usage{
vjpair.group.plot(
  group1_clustered_seqs,
  group1_label,
  group2_clustered_seqs,
  group2_label
)
}
\arguments{
\item{group1_clustered_seqs}{Clustered sequences from group1}

\item{group1_label}{Label of group1}

\item{group2_clustered_seqs}{Clustered sequences from group2}

\item{group2_label}{Label of group2}
}
\value{
Heatmap
}
\description{
Plot: Heatmap of VJpair gene usage frequency between groups
}
