% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream_Analysis.R
\name{vjpair.sample.plot}
\alias{vjpair.sample.plot}
\title{Plot: Heatmap of VJpair gene usage frequency within a sample}
\usage{
vjpair.sample.plot(clustered_seqs)
}
\arguments{
\item{clustered_seqs}{Clustered sequences from a sample}
}
\value{
Heatmap
}
\description{
Plot: Heatmap of VJpair gene usage frequency within a sample
}
